feat: extracted computeType, createJsYamlSchema, createTypeNode into their own fiels
