+++ packages/gatsby-transformer-raml/gatsby-node.js
@@ -1,5 +1,7 @@
 const firstline = require('firstline');
 const jsYaml = require('js-yaml');
+const createTypeNode = require('./src/create-type-node');
+const createJsYamlSchema = require('./src/create-js-yaml-schema');
 
 async function onCreateNode({
   node,
@@ -35,125 +37,6 @@ async function onCreateNode({
       reporter.error(e);
     }
   }
-}
-
-function createJsYamlSchema() {
-  /**
-   * Without definition of custom types, js-yaml throws an error when
-   * it hits tags like "!include".
-   *
-   * See here for more info on how to write custom types for js-yaml
-   * https://github.com/nodeca/js-yaml/wiki/Custom-types
-   */
-  const IncludeYamlType = new jsYaml.Type('!include', {
-    kind: 'scalar',
-    construct: data => {
-      return data !== null ? data : '';
-    },
-  });
-
-  return jsYaml.Schema.create([IncludeYamlType]);
-}
-
-function createTypeNode({
-  type,
-  fileNode,
-  createNode,
-  createNodeId,
-  createParentChildLink,
-  createContentDigest,
-}) {
-  const postProcessedType = postProcessType(type, fileNode.relativeDirectory);
-
-  const typeNode = {
-    ...postProcessedType,
-    id: createNodeId(`${fileNode.id} >>> RAML_TYPE`),
-    children: [],
-    parent: fileNode.id,
-    internal: {
-      contentDigest: createContentDigest(postProcessedType),
-      mediaType: fileNode.internal.mediaType,
-      type: 'RamlType',
-    },
-  };
-
-  createNode(typeNode);
-  createParentChildLink({ parent: fileNode, child: typeNode });
-}
-
-function postProcessType(type, fileNodeRelativeDirectory) {
-  const postProcessedType = extractParenthesisFromAnnotationIdentifier(type);
-
-  postProcessedType.apiKey = fileNodeRelativeDirectory.replace(`/types`, '');
-  postProcessedType.properties = propertiesToArrays(
-    postProcessedType.properties
-  );
-  postProcessedType.examples = examplesToArrays(postProcessedType.examples);
-  postProcessedType.enumDescriptions = enumDescriptionsToArray(
-    postProcessedType.enumDescriptions
-  );
-
-  return postProcessedType;
-}
-
-function extractParenthesisFromAnnotationIdentifier(type) {
-  const returnedType = {};
-
-  Object.keys(type).forEach(key => {
-    const keyWithoutParenthesis = key.replace(`(`, '').replace(`)`, '');
-
-    if (computeType(type[key]) === 'object') {
-      returnedType[
-        keyWithoutParenthesis
-      ] = extractParenthesisFromAnnotationIdentifier(type[key]);
-      return;
-    }
-
-    returnedType[keyWithoutParenthesis] = type[key];
-  });
-
-  return returnedType;
-}
-
-function computeType(value) {
-  /**
-   * More info about why we need this better type checker here:
-   * https://blog.logrocket.com/javascript-typeof-2511d53a1a62/
-   */
-  const regex = /^\[object (\S+?)\]$/;
-  const matches = Object.prototype.toString.call(value).match(regex) || [];
-
-  return (matches[1] || 'undefined').toLowerCase();
-}
-
-function propertiesToArrays(properties) {
-  if (properties) {
-    return Object.entries(properties).map(([key, value]) => {
-      return { ...value, name: key };
-    });
-  }
-
-  return undefined;
-}
-
-function examplesToArrays(examples) {
-  if (examples) {
-    return Object.entries(examples).map(([key, value]) => {
-      return { name: key, file: value };
-    });
-  }
-
-  return undefined;
-}
-
-function enumDescriptionsToArray(enumDescriptions) {
-  if (enumDescriptions) {
-    return Object.entries(enumDescriptions).map(([key, value]) => {
-      return { name: key, description: value };
-    });
-  }
-
-  return undefined;
 }
 
 exports.onCreateNode = onCreateNode;
