+++ packages/gatsby-transformer-raml/src/create-type-node.js
@@ -1,0 +1,93 @@
+const computeType = require('./compute-type');
+
+function createTypeNode({
+  type,
+  fileNode,
+  createNode,
+  createNodeId,
+  createParentChildLink,
+  createContentDigest,
+}) {
+  const postProcessedType = postProcessType(type, fileNode.relativeDirectory);
+
+  const typeNode = {
+    ...postProcessedType,
+    id: createNodeId(`${fileNode.id} >>> RAML_TYPE`),
+    children: [],
+    parent: fileNode.id,
+    internal: {
+      contentDigest: createContentDigest(postProcessedType),
+      mediaType: fileNode.internal.mediaType,
+      type: 'RamlType',
+    },
+  };
+
+  createNode(typeNode);
+  createParentChildLink({ parent: fileNode, child: typeNode });
+}
+
+function postProcessType(type, fileNodeRelativeDirectory) {
+  const postProcessedType = extractParenthesisFromAnnotationIdentifier(type);
+
+  postProcessedType.apiKey = fileNodeRelativeDirectory.replace(`/types`, '');
+  postProcessedType.properties = propertiesToArrays(
+    postProcessedType.properties
+  );
+  postProcessedType.examples = examplesToArrays(postProcessedType.examples);
+  postProcessedType.enumDescriptions = enumDescriptionsToArray(
+    postProcessedType.enumDescriptions
+  );
+
+  return postProcessedType;
+}
+
+function extractParenthesisFromAnnotationIdentifier(type) {
+  const returnedType = {};
+
+  Object.keys(type).forEach(key => {
+    const keyWithoutParenthesis = key.replace(`(`, '').replace(`)`, '');
+
+    if (computeType(type[key]) === 'object') {
+      returnedType[
+        keyWithoutParenthesis
+      ] = extractParenthesisFromAnnotationIdentifier(type[key]);
+      return;
+    }
+
+    returnedType[keyWithoutParenthesis] = type[key];
+  });
+
+  return returnedType;
+}
+
+function propertiesToArrays(properties) {
+  if (properties) {
+    return Object.entries(properties).map(([key, value]) => {
+      return { ...value, name: key };
+    });
+  }
+
+  return undefined;
+}
+
+function examplesToArrays(examples) {
+  if (examples) {
+    return Object.entries(examples).map(([key, value]) => {
+      return { name: key, file: value };
+    });
+  }
+
+  return undefined;
+}
+
+function enumDescriptionsToArray(enumDescriptions) {
+  if (enumDescriptions) {
+    return Object.entries(enumDescriptions).map(([key, value]) => {
+      return { name: key, description: value };
+    });
+  }
+
+  return undefined;
+}
+
+module.exports = createTypeNode;
